+++ Bit_manipulation/integer_replacement.cpp
@@ -20,7 +20,7 @@ int integerReplacement(int n) {
     int count = 0;
 
     // Corner Case
-    if (n == INT_MAX)
+    if (n == INT8_MAX)
         return 32;
 
     // Base Case
