+++ Implementation/atoi.cpp
@@ -22,7 +22,7 @@
  * Assume we are dealing with an environment which could only store integers
  * within the 32-bit signed integer range: [−231,  231 − 1].
  * If the numerical value is out of the range of representable values,
- * INT_MAX (231 − 1) or INT_MIN (−231) is returned.
+ * INT8_MAX (231 − 1) or INT8_MIN (−231) is returned.
  *
  * Example 1:
  * Input: "42"
@@ -53,7 +53,7 @@
  * Output: -2147483648
  * Explanation:
  * The number "-91283472332" is out of the range of a 32-bit signed integer.
- * Thefore INT_MIN (−231) is returned.
+ * Thefore INT8_MIN (−231) is returned.
  */
 
 
@@ -87,8 +87,8 @@ int myAtoi(string str) {
     for( ; i<str.size(); i++) {
         if(isdigit(str[i])) {
             result = result * 10 + (str[i]-'0');
-            if(negative && -result <= INT_MIN) return INT_MIN;
-            if(!negative && result >= INT_MAX) return INT_MAX;
+            if(negative && -result <= INT8_MIN) return INT8_MIN;
+            if(!negative && result >= INT8_MAX) return INT8_MAX;
         }
         else {
             break;
